Restructure preflight and get context from global flag --context
