+++ pkg/cmd/init/preflight/testdata/kubeconfig
@@ -2,7 +2,7 @@ apiVersion: v1
 kind: Config
 clusters:
 - cluster:
-    server: https://1.2.3.4:8443
+    server: https://localhost:8443
   name: cluster-hub
 contexts:
 - context:
