+++ pkg/genericclioptions/clusteradm_flags.go
@@ -24,5 +24,11 @@ func NewClusteradmFlags(f cmdutil.Factory) *ClusteradmFlags {
 func (f *ClusteradmFlags) AddFlags(flags *pflag.FlagSet) {
 	flags.BoolVar(&f.DryRun, "dry-run", false, "If set the generated resources will be displayed but not applied")
 	flags.IntVar(&f.Timeout, "timeout", 300, "extend timeout from 300 secounds ")
-	flags.StringVar(&f.Context, "context", "", "The name of the kubeconfig context to use")
+}
+
+// SetContext will set current context from command line argument --context.
+func (f *ClusteradmFlags) SetContext(context *string) {
+	if context != nil {
+		f.Context = *context
+	}
 }
